Rename issue_310 to issue_632
Use Github issue numbers.
