+++ air/tests/test_module/issues/mod.rs
@@ -31,9 +31,9 @@ mod issue_300;
 mod issue_302;
 mod issue_304;
 mod issue_306;
-mod issue_310;
 mod issue_331;
 mod issue_346;
 mod issue_348;
 mod issue_356;
 mod issue_363;
+mod issue_632;
